requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
pysam>=0.21.0
biopython==1.83
matplotlib>=3.7.0
seaborn>=0.12.0
pyinstaller==6.3.0
